Move over some more common code
