+++ link-grammar/dict-common/dict-internals.c
@@ -13,13 +13,26 @@
 
 #include "dict-internals.h"
 
+void dict_node_noop(Dictionary dict)
+{
+}
+
+Dict_node * dict_node_new(void)
+{
+   return (Dict_node*) malloc(sizeof(Dict_node));
+}
+
+void dict_node_free_list(Dict_node *llist)
+{
+   while (llist != NULL)
+   {
+      Dict_node * n = llist->right;
+      free(llist);
+      llist = n;
+   }
+}
+
 void dict_node_free_lookup(Dictionary dict, Dict_node* llist)
 {
-	Dict_node * dn;
-	while (llist != NULL)
-	{
-		dn = llist->right;
-		free(llist);
-		llist = dn;
-	}
+	dict_node_free_list(llist);
 }
