+++ link-grammar/dict-ram/dict-ram.h
@@ -10,7 +10,6 @@
 /*                                                                       */
 /*************************************************************************/
 
-Dict_node * dict_node_new(void);
 Dict_node * strict_lookup_list(const Dictionary dict, const char *s);
 Dict_node * dsw_tree_to_vine (Dict_node *root);
 Dict_node * dsw_vine_to_tree (Dict_node *root, int size);
@@ -19,10 +18,8 @@ Dict_node * dict_node_insert(Dictionary dict, Dict_node *n, Dict_node *newnode);
 Dict_node * dict_node_lookup(const Dictionary dict, const char *s);
 Dict_node * dict_node_wild_lookup(Dictionary dict, const char *s);
 bool dict_node_exists_lookup(Dictionary dict, const char *s);
-void dict_node_free_list(Dict_node *llist);
 
 void free_dictionary_root(Dictionary dict);
-void dict_node_noop(Dictionary dict);
 
 Exp * Exp_create(Pool_desc *);
 Exp * Exp_create_dup(Pool_desc *, Exp *);
